Move the types package to the compiler package.
