+++ internal/vm/compiler/ast/ast.go
@@ -10,7 +10,7 @@ import (
 	"github.com/google/mtail/internal/metrics"
 	"github.com/google/mtail/internal/vm/compiler/position"
 	"github.com/google/mtail/internal/vm/compiler/symbol"
-	"github.com/google/mtail/internal/vm/types"
+	"github.com/google/mtail/internal/vm/compiler/types"
 )
 
 type Node interface {
