+++ internal/vm/compiler/checker/checker.go
@@ -14,7 +14,7 @@ import (
 	"github.com/google/mtail/internal/vm/compiler/errors"
 	"github.com/google/mtail/internal/vm/compiler/parser"
 	"github.com/google/mtail/internal/vm/compiler/symbol"
-	"github.com/google/mtail/internal/vm/types"
+	"github.com/google/mtail/internal/vm/compiler/types"
 )
 
 const (
