+++ internal/vm/compiler/types/regexp.go
